[CLN] sale_loyalty: extract tests on rewards
