+++ addons/sale_loyalty/tests/__init__.py
@@ -3,6 +3,7 @@
 from . import test_buy_gift_card
 from . import test_loyalty
 from . import test_loyalty_history
+from . import test_loyalty_reward
 from . import test_pay_with_gift_card
 from . import test_program_multi_company
 from . import test_program_numbers
